#110 added filter suite UI test and removed individual filter UI tests
