--- app/src/androidTest/java/com/simplemobiletools/camera/activities/CameraBurstTest.kt
@@ -1,75 +1,0 @@
-package com.simplemobiletools.camera.activities
-
-import android.view.View
-import android.view.ViewGroup
-import androidx.test.espresso.Espresso
-import androidx.test.espresso.action.ViewActions
-import androidx.test.espresso.matcher.ViewMatchers
-import androidx.test.filters.LargeTest
-import androidx.test.rule.ActivityTestRule
-import androidx.test.rule.GrantPermissionRule
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.simplemobiletools.camera.R
-import com.simplemobiletools.commons.extensions.isVisible
-import kotlinx.android.synthetic.main.activity_main.*
-import org.hamcrest.Description
-import org.hamcrest.Matcher
-import org.hamcrest.Matchers
-import org.hamcrest.TypeSafeMatcher
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-
-@LargeTest
-@RunWith(AndroidJUnit4::class)
-class BurstShotUITest {
-
-    @Rule
-    @JvmField
-    var mActivityTestRule = ActivityTestRule(MainActivity::class.java)
-
-    @Rule
-    @JvmField
-    var mGrantPermissionRule =
-            GrantPermissionRule.grant(
-                    "android.permission.CAMERA",
-                    "android.permission.WRITE_EXTERNAL_STORAGE")
-
-    @Test
-    fun burstShotUITest() {
-        // Added a sleep statement to match the app's execution delay.
-        // The recommended way to handle such scenarios is to use Espresso idling resources:
-        // https://google.github.io/android-testing-support-library/docs/espresso/idling-resource/index.html
-        Thread.sleep(1000)
-
-        val appCompatImageView = Espresso.onView(
-
-                Matchers.allOf(ViewMatchers.withId(R.id.shutter),
-                        childAtPosition(
-                                Matchers.allOf(ViewMatchers.withId(R.id.btn_holder),
-                                        childAtPosition(
-                                                ViewMatchers.withId(R.id.view_holder),
-                                                5)),
-                                1),
-                        ViewMatchers.isDisplayed()))
-
-        appCompatImageView.perform(ViewActions.longClick())
-    }
-
-    private fun childAtPosition(
-            parentMatcher: Matcher<View>, position: Int): Matcher<View> {
-
-        return object : TypeSafeMatcher<View>() {
-            override fun describeTo(description: Description) {
-                description.appendText("Child at position $position in parent ")
-                parentMatcher.describeTo(description)
-            }
-
-            public override fun matchesSafely(view: View): Boolean {
-                val parent = view.parent
-                return parent is ViewGroup && parentMatcher.matches(parent)
-                        && view == parent.getChildAt(position)
-            }
-        }
-    }
-}
