--- app/src/androidTest/java/com/simplemobiletools/camera/activities/InvertFilterTest.kt
@@ -1,103 +1,0 @@
-package com.simplemobiletools.camera.activities
-
-
-import android.hardware.camera2.CameraMetadata
-import android.view.View
-import android.view.ViewGroup
-import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.action.ViewActions.click
-import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
-import androidx.test.espresso.matcher.ViewMatchers.withId
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.filters.LargeTest
-import androidx.test.rule.ActivityTestRule
-import androidx.test.rule.GrantPermissionRule
-import com.simplemobiletools.camera.R
-import com.simplemobiletools.camera.views.CameraPreview
-import org.junit.Assert
-import org.hamcrest.Description
-import org.hamcrest.Matcher
-import org.hamcrest.Matchers.allOf
-import org.hamcrest.TypeSafeMatcher
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-
-@LargeTest
-@RunWith(AndroidJUnit4::class)
-class InvertFilterTest {
-
-    @Rule
-    @JvmField
-    var mActivityTestRule = ActivityTestRule(SplashActivity::class.java)
-
-    @Rule
-    @JvmField
-    var mainActivityTestRule = ActivityTestRule(MainActivity::class.java)
-
-    @Rule
-    @JvmField
-    var mGrantPermissionRule =
-            GrantPermissionRule.grant(
-                    "android.permission.CAMERA",
-                    "android.permission.WRITE_EXTERNAL_STORAGE")
-
-    @Test
-    fun invertFilterTest() {
-        // Added a sleep statement to match the app's execution delay.
-        // The recommended way to handle such scenarios is to use Espresso idling resources:
-        // https://google.github.io/android-testing-support-library/docs/espresso/idling-resource/index.html
-        Thread.sleep(5000)
-
-        val appCompatImageView = onView(
-                allOf(withId(R.id.filter),
-                        childAtPosition(
-                                allOf(withId(R.id.view_holder),
-                                        childAtPosition(
-                                                withId(android.R.id.content),
-                                                0)),
-                                1),
-                        isDisplayed()))
-        appCompatImageView.perform(click())
-
-        val appCompatImageView2 = onView(
-                allOf(withId(R.id.invert),
-                        childAtPosition(
-                                allOf(withId(R.id.view_holder),
-                                        childAtPosition(
-                                                withId(android.R.id.content),
-                                                0)),
-                                2),
-                        isDisplayed()))
-        appCompatImageView2.perform(click())
-    }
-
-    @Test
-    fun testInvertFilter() {
-
-        val testedPreview = (mainActivityTestRule.activity.getMPreview()) as CameraPreview
-
-        Assert.assertEquals(testedPreview.getMCameraEffect(), CameraMetadata.CONTROL_EFFECT_MODE_OFF)
-
-        testedPreview.setCameraEffect("invert")
-
-        Assert.assertEquals(testedPreview.getMCameraEffect(), CameraMetadata.CONTROL_EFFECT_MODE_NEGATIVE)
-    }
-
-    private fun childAtPosition(
-            parentMatcher: Matcher<View>, position: Int): Matcher<View> {
-
-        return object : TypeSafeMatcher<View>() {
-            override fun describeTo(description: Description) {
-                description.appendText("Child at position $position in parent ")
-                parentMatcher.describeTo(description)
-            }
-
-            public override fun matchesSafely(view: View): Boolean {
-                val parent = view.parent
-                return parent is ViewGroup && parentMatcher.matches(parent)
-                        && view == parent.getChildAt(position)
-            }
-        }
-    }
-}
